+++ crates/sncast/tests/e2e/script/mod.rs
@@ -1,7 +1,0 @@
-mod call;
-mod declare;
-mod deploy;
-mod general;
-mod init;
-mod invoke;
-mod tx_status;